@@ -3,11 +3,10 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "dev": "craco start",
-    "start": "craco start",
-    "build": "craco build && node ./scripts/writeRedirects.js",
-    "test": "craco test",
-    "eject": "react-scripts eject",
+    "dev": "next dev",
+    "start": "next start",
+    "build": "next build",
+    "lint": "next lint",
     "generate": "graphql-codegen",
     "contracts": "node ./scripts/contracts.js",
     "evm": "cd ../core-contracts && yarn evm",
@@ -16,28 +15,27 @@
   "dependencies": {
     "@0xsequence/multicall": "^0.28.0",
     "@apollo/client": "^3.4.7",
-    "@craco/craco": "^6.1.2",
     "@dopex-io/sdk": "../sdk",
     "@material-ui/core": "^4.11.2",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.58",
     "@walletconnect/web3-provider": "^1.6.5",
-    "axios": "^0.21.1",
+    "axios": "^0.24.0",
     "bignumber.js": "^9.0.1",
     "classnames": "^2.2.6",
     "date-fns": "^2.16.1",
     "eth-rpc-errors": "^4.0.2",
-    "ethers": "^5.4.5",
+    "ethers": "5.4.7",
     "formik": "^2.2.6",
     "fortmatic": "^2.2.1",
-    "graphql": "^15.5.1",
+    "graphql": "^16.0.1",
     "lodash": "^4.17.20",
+    "next": "^12.0.3",
     "react": "^17.0.1",
     "react-countdown": "^2.3.2",
     "react-dom": "^17.0.1",
     "react-hot-toast": "^2.0.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
+    "react-router-dom": "^6.0.2",
     "react-table": "^7.7.0",
     "react-use": "^17.2.4",
     "recharts": "^2.0.9",
@@ -50,27 +48,22 @@
     "@graphql-codegen/typescript": "^2.0.0",
     "@graphql-codegen/typescript-operations": "^2.0.1",
     "@graphql-codegen/typescript-react-apollo": "^3.0.0",
+    "@next/eslint-plugin-next": "^12.0.3",
     "@types/lodash": "^4.14.170",
     "@types/react": "^17.0.11",
     "@types/react-dom": "^17.0.8",
     "@types/react-router-dom": "^5.1.7",
     "@typescript-eslint/eslint-plugin": "^4.0.0",
     "@typescript-eslint/parser": "^4.0.0",
-    "autoprefixer": "^9",
-    "babel-eslint": "^10.0.0",
-    "eslint": "^7.5.0",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-flowtype": "^5.2.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-react": "^7.20.3",
-    "eslint-plugin-react-hooks": "^4.0.8",
+    "autoprefixer": "^10.4.0",
+    "eslint": "<8.0.0",
+    "eslint-config-next": "12.0.3",
     "husky": "^7.0.1",
-    "lint-staged": "^11.0.0",
-    "postcss": "7.0.36",
+    "lint-staged": "^12.0.1",
+    "postcss": "^8.3.11",
     "prettier": "^2.1.1",
     "shelljs": "^0.8.4",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat",
+    "tailwindcss": "^2.2.19",
     "typescript": "^4.1.2"
   },
   "lint-staged": {
