@@ -1,7 +1,11 @@
 import { Suspense, lazy } from 'react';
-import { BrowserRouter, Route, Switch, Redirect } from 'react-router-dom';
+import { BrowserRouter, Route, Routes, Navigate } from 'react-router-dom';
 import { Toaster } from 'react-hot-toast';
 import { ApolloProvider } from '@apollo/client';
+import { ThemeProvider, StylesProvider } from '@material-ui/core/styles';
+import Error from 'next/error';
+
+import theme from './style/muiTheme';
 
 import { client } from 'graphql/apollo';
 
@@ -15,11 +19,11 @@ import { SsovProvider } from 'contexts/Ssov';
 import ChangeNetworkDialog from 'components/ChangeNetworkDialog';
 import PageLoader from 'components/PageLoader';
 
-const Farming = lazy(() => import('pages/farming/farms'));
-const FarmingStake = lazy(() => import('pages/farming/stake'));
-const TokenSale = lazy(() => import('pages/sale'));
-const Ssov = lazy(() => import('pages/ssov'));
-const SsovManage = lazy(() => import('pages/ssov/Manage'));
+const Farming = lazy(() => import('craPages/farming/farms'));
+const FarmingStake = lazy(() => import('craPages/farming/stake'));
+const TokenSale = lazy(() => import('craPages/sale'));
+const Ssov = lazy(() => import('craPages/ssov'));
+const SsovManage = lazy(() => import('craPages/ssov/Manage'));
 // const Portfolio = lazy(() => import('pages/portfolio'));
 // const Options = lazy(() => import('pages/options'));
 // const Pools = lazy(() => import('pages/pools'));
@@ -27,6 +31,28 @@ const SsovManage = lazy(() => import('pages/ssov/Manage'));
 // const PoolsVolume = lazy(() => import('pages/pools/volume'));
 // const TestnetFaucet = lazy(() => import('pages/testnet-faucet'));
 // const Swap = lazy(() => import('pages/swap'));
+
+const FarmRoutes = () => {
+  return (
+    <FarmingProvider>
+      <Routes>
+        <Route path="*" element={<Farming />} />
+        <Route path="stake" element={<FarmingStake />} />
+      </Routes>
+    </FarmingProvider>
+  );
+};
+
+const SsovRoutes = () => {
+  return (
+    <SsovProvider>
+      <Routes>
+        <Route path="*" element={<Ssov />} />
+        <Route path="manage/:asset" element={<SsovManage />} />
+      </Routes>
+    </SsovProvider>
+  );
+};
 
 function AppRoutes() {
   // if (BUILD === 'testnet') {
@@ -51,28 +77,17 @@ function AppRoutes() {
   //   );
   // }
   return (
-    <BrowserRouter forceRefresh={false}>
+    <BrowserRouter>
       <Suspense fallback={<PageLoader />}>
         <WalletProvider>
           <AssetsProvider>
-            <FarmingProvider>
-              <SsovProvider>
-                <Switch>
-                  <Route path="/" exact>
-                    <Redirect to="/ssov" />
-                  </Route>
-                  <Route path="/sale" component={TokenSale} exact />
-                  <Route path="/ssov" component={Ssov} exact />
-                  <Route
-                    path="/ssov/manage/:asset"
-                    component={SsovManage}
-                    exact
-                  />
-                  <Route path="/farms" component={Farming} exact />
-                  <Route path="/farms/stake" component={FarmingStake} exact />
-                </Switch>
-              </SsovProvider>
-            </FarmingProvider>
+            <Routes>
+              <Route path="/" element={<Navigate to="/ssov" />} />
+              <Route path="sale" element={<TokenSale />} />
+              <Route path="ssov/*" element={<SsovRoutes />} />
+              <Route path="farms/*" element={<FarmRoutes />} />
+              <Route path="*" element={<Error statusCode={404} />} />
+            </Routes>
             <ChangeNetworkDialog />
           </AssetsProvider>
         </WalletProvider>
@@ -83,10 +98,14 @@ function AppRoutes() {
 
 const App = () => {
   return (
-    <ApolloProvider client={client}>
-      <Toaster position="bottom-right" reverseOrder={true} />
-      <AppRoutes />
-    </ApolloProvider>
+    <StylesProvider injectFirst>
+      <ThemeProvider theme={theme}>
+        <ApolloProvider client={client}>
+          <Toaster position="bottom-right" reverseOrder={true} />
+          <AppRoutes />
+        </ApolloProvider>
+      </ThemeProvider>
+    </StylesProvider>
   );
 };
 
